PHP7.2 compatibility. BC compatible. (#852)
* PHP7.2 compatibility. BC compatible.

* fix cs

* set unlimited memory_limit

* add memory limit before composer.

* debug

* fix

* fix tests on sf2

* travis. remove php -i

* fix typo.

* fix ES endpoint in tests.

* remove 3.5, 3.6 symfony from travis matrix.

* use yaml extension for sf4 config.

* trigger travis

* trigger travis.
