+++ .travis.yml
@@ -15,16 +15,16 @@ env:
     - SYMFONY="~3.2"
     - SYMFONY="~3.3"
     - SYMFONY="~3.4"
-    - SYMFONY="~3.5"
-    - SYMFONY="~3.6"
 install:
   - wget ${ES_DOWNLOAD_URL}
   - tar -xzf elasticsearch-${ES_VERSION}.tar.gz
   - ./elasticsearch-${ES_VERSION}/bin/elasticsearch -d
 before_script:
+  - echo 'memory_limit=-1' >> ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d/travis.ini
   - composer require --no-update symfony/symfony:${SYMFONY}
   - composer config -g github-oauth.github.com $GITHUB_COMPOSER_AUTH
   - composer install --no-interaction --prefer-dist
+
 script:
   - wget -q --waitretry=1 --retry-connrefused -T 10 -O - http://127.0.0.1:9200
   - vendor/bin/phpunit --coverage-clover=coverage.clover
