+++ Resources/config/services.yml
@@ -2,11 +2,6 @@ parameters:
     es.logging.path: "%kernel.logs_dir%/elasticsearch_%kernel.environment%.log"
 
 services:
-    ONGR\ElasticsearchBundle\Command\:
-        resource: '../../Command'
-        tags:
-            - { name: console.command }
-
     es.export:
         class: ONGR\ElasticsearchBundle\Service\ExportService
