+++ Result/Converter.php
@@ -13,7 +13,7 @@ namespace ONGR\ElasticsearchBundle\Result;
 
 use Doctrine\Common\Collections\Collection;
 use ONGR\ElasticsearchBundle\Annotation\Nested;
-use ONGR\ElasticsearchBundle\Annotation\Object;
+use ONGR\ElasticsearchBundle\Annotation\ObjectType;
 use ONGR\ElasticsearchBundle\Mapping\MetadataCollector;
 use ONGR\ElasticsearchBundle\Service\Manager;
 
@@ -104,7 +104,7 @@ class Converter
                             $value = new \DateTime($value);
                         }
                         break;
-                    case Object::NAME:
+                    case ObjectType::NAME:
                     case Nested::NAME:
                         if ($aliases[$name]['multiple']) {
                             $value = new ObjectIterator($this, $value, $aliases[$name]);
